Bump Ecto timeout
We're currently timing out on large playlists. When we update the
playlist_items list for a given playlist, we use put_assoc, which
performs the update/insert/deletions in a single transaction. If that
step isn't completed before the timeout, `put_assoc` will raise an
exception and the import will fail.

To fix, bump the timeout to a more reasonable amount and re-evaluate if
necessary.